Combine test files into single test_curate_and_extract_coding_seqs.py
Merged test_create_unaligned_coding_seqs.py and test_curate_msa.py into
a single test file since they both test functions from the same script
(curate_and_extract_coding_seqs.py). All 94 tests pass successfully.
